page account coding

+++ dao/src/main/resources/db/migration/V1__init_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE t_account
 (
-    id          int8        NOT NULL,
+    id          bigint        NOT NULL,
     create_time timestamp   NOT NULL,
     update_time timestamp   NOT NULL,
     pwd  varchar(64) NOT NULL,
@@ -10,7 +10,7 @@ CREATE TABLE t_account
 
 CREATE TABLE t_role
 (
-    id          int8        NOT NULL,
+    id          bigint        NOT NULL,
     create_time timestamp   NOT NULL,
     update_time timestamp   NOT NULL,
     role_name      varchar(64) NOT NULL,
@@ -19,7 +19,7 @@ CREATE TABLE t_role
 
 CREATE TABLE t_account_role
 (
-    account_id int8 NOT NULL,
-    role_id    int8 NOT NULL,
+    account_id bigint NOT NULL,
+    role_id    bigint NOT NULL,
     CONSTRAINT t_account_role_pkey PRIMARY KEY (account_id, role_id)
 );
